# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API document",
        "version": "1.0.0"
    },
    "paths": {
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "description": "Returns list of events",
                "operationId": "e529c81e6aaec82f49950a87ed881f57",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/events/{event}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "description": "Get a single event from the ID",
                "operationId": "2f5e4d4d8d7989ca1f72be9c20d9daac",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get event susscess"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/api/events/{event}/editable": {
            "post": {
                "tags": [
                    "Events"
                ],
                "description": "Check if that event is still editable",
                "operationId": "8805e3b238c89ea1b4f5807f19423cc8",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Allowed edit event"
                    },
                    "409": {
                        "description": "conflict"
                    }
                }
            }
        },
        "/api/events/{event}/editable/release": {
            "post": {
                "tags": [
                    "Events"
                ],
                "description": "Release event",
                "operationId": "7712aa46adaea6908d2cb690d54a862f",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Release event successfully"
                    }
                }
            }
        },
        "/api/events/{event}/editable/maintain": {
            "post": {
                "tags": [
                    "Events"
                ],
                "description": "Maintain event",
                "operationId": "26747794bb22e08e6664349160fb7f94",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Maintain success"
                    },
                    "409": {
                        "description": "conflict"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "535ffeda4b78916efd658b2844ea897a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "admin@test",
                                    "password": 12345678
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Returns my info",
                "operationId": "48337fc91254ad3a6712be5cacb7c483",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Logout",
                "operationId": "46f8a14c8083abb69eed72768aa08629",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Refresh token",
                "operationId": "f3538b6e9c46bcf601a9d194288b03d4",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}